gen-ai/lang-chain/requirements.txt
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-experimental>=0.3,<0.4
langchain-openai>=0.0.5
python-dotenv>=0.21.0

# DuckDuckGo
duckduckgo-search>=3.0.0

# Yahoo Finance
yfinance>=0.2.36